+++ application/common/models/User.php
@@ -10,13 +10,14 @@ class User
     const EMAIL = 'email';
     const EMPLOYEE_ID = 'employee_id';
     const FIRST_NAME = 'first_name';
+    const GROUPS = 'groups';
     const LAST_NAME = 'last_name';
     const LOCKED = 'locked';
     const MANAGER_EMAIL = 'manager_email';
+    const PERSONAL_EMAIL = 'personal_email';
     const REQUIRE_MFA = 'require_mfa';
     const SPOUSE_EMAIL = 'spouse_email';
     const USERNAME = 'username';
-    const GROUPS = 'groups';
 
     /**
      * The values (indexed by field name) for the fields which have been set.
@@ -64,13 +65,14 @@ class User
             self::EMAIL,
             self::EMPLOYEE_ID,
             self::FIRST_NAME,
+            self::GROUPS,
             self::LAST_NAME,
             self::LOCKED,
             self::MANAGER_EMAIL,
+            self::PERSONAL_EMAIL,
             self::REQUIRE_MFA,
             self::SPOUSE_EMAIL,
             self::USERNAME,
-            self::GROUPS
         ];
     }
     
@@ -160,6 +162,14 @@ class User
     public function getSpouseEmail()
     {
         return $this->values[self::SPOUSE_EMAIL] ?? null;
+    }
+
+    /**
+     * @return null|string
+     */
+    public function getPersonalEmail()
+    {
+        return $this->values[self::PERSONAL_EMAIL] ?? null;
     }
 
     /**